+++ unlock-js/src/v11/purchaseKey.js
@@ -1,8 +1,8 @@
 import utils from '../utils'
-import { GAS_AMOUNTS } from '../constants'
+import { GAS_AMOUNTS, ZERO } from '../constants'
 import TransactionTypes from '../transactionTypes'
 import Errors from '../errors'
-import { approveTransfer } from '../erc20'
+import { approveTransfer, getErc20Decimals } from '../erc20'
 
 /**
  * Purchase key function. This implementation requires the following
@@ -12,15 +12,30 @@ import { approveTransfer } from '../erc20'
  * - {string} keyPrice
  * - {PropTypes.address} erc20Address
  * - {number} decimals
+ * @return {string} hash of the transaction
  */
 export default async function({
   lockAddress,
   owner,
   keyPrice,
   erc20Address,
-  decimals = 18,
+  decimals,
 }) {
   const lockContract = await this.getLockContract(lockAddress)
+
+  if (!erc20Address || erc20Address !== ZERO) {
+    erc20Address = await lockContract.tokenAddress()
+  }
+
+  // decimals could be 0!
+  if (decimals == null) {
+    // get the decimals from the ERC20 contract or default to 18
+    if (erc20Address && erc20Address !== ZERO) {
+      decimals = await getErc20Decimals(erc20Address, this.provider)
+    } else {
+      decimals = 18
+    }
+  }
 
   const actualAmount = utils.toDecimal(keyPrice, decimals)
 
@@ -28,7 +43,7 @@ export default async function({
     gasLimit: GAS_AMOUNTS.purchaseFor,
   }
 
-  if (erc20Address) {
+  if (erc20Address && erc20Address !== ZERO) {
     await approveTransfer(
       erc20Address,
       lockAddress,
@@ -45,11 +60,10 @@ export default async function({
       owner,
       purchaseForOptions
     )
-    const ret = await this._handleMethodCall(
+    return await this._handleMethodCall(
       transactionPromise,
       TransactionTypes.KEY_PURCHASE
     )
-    return ret
   } catch (error) {
     this.emit('error', new Error(Errors.FAILED_TO_PURCHASE_KEY))
   }