+++ unlock-js/src/walletService.js
@@ -120,7 +120,7 @@ export default class WalletService extends UnlockService {
       transactionType,
       'submitted'
     )
-    const finalTransaction = await transaction.wait()
+    const finalTransaction = await transaction.wait() // TODO: why do we wait here? This should return instantly: getting a hash should not require i/o
     return finalTransaction.hash
     // errors fall through
   }